pyproject.toml
[project]
name = "package-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kevin Cash", email = "kevincash123@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "rich>=14.3.1",
    "typer>=0.12.0",
]

[project.scripts]
package-downloader = "package_downloader:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
